Grammar
It can handle: Global variable declaration, Basic FOR loop, WHILE loop, IF statement, and Nested if statement. Block can contain declaration, assignment, print, addition, subtraction and less than, greater than, or equal operations. According to the grammar, it is not allowed to declare a variable and assign it a value in the same line.